fix: Move ConsensusConstants to consensus crate
