+++ base/src/consensus_state.rs
@@ -4,7 +4,7 @@
 use crate::{
     epoch_data::EpochData,
     global_slot::GlobalSlot,
-    numbers::{Amount, GlobalSlotNumber, Length},
+    numbers::{Amount, BlockTime, GlobalSlotNumber, Length},
 };
 
 use mina_crypto::{base58::Base58Encodable, hash::*, signature::PublicKey};
@@ -74,6 +74,38 @@ pub struct ConsensusState {
     pub coinbase_receiver: PublicKey,
     /// true if block_stake_winner has no locked tokens, false otherwise
     pub supercharge_coinbase: bool,
+}
+
+pub struct ConsensusConstants {
+    /// Point of finality (number of confirmations)
+    pub k: Length,
+    /// Number of slots per epoch
+    pub slots_per_epoch: Length,
+    /// No of slots in a sub-window = 7
+    pub slots_per_sub_window: Length,
+    /// Maximum permissable delay of packets (in slots after the current)
+    pub delta: Length,
+    /// Timestamp of genesis block in unixtime
+    pub genesis_state_timestamp: BlockTime,
+    /// Sub windows within a window
+    pub sub_windows_per_window: Length,
+    /// Number of slots before minimum density is used in chain selection
+    pub grace_period_end: Length,
+}
+
+impl Default for ConsensusConstants {
+    fn default() -> Self {
+        // TODO: read from config
+        Self {
+            k: Length(290),
+            slots_per_epoch: 7140.into(),
+            slots_per_sub_window: 7.into(),
+            delta: 0.into(),
+            genesis_state_timestamp: BlockTime(1615939200000),
+            sub_windows_per_window: 11.into(),
+            grace_period_end: Length(1440),
+        }
+    }
 }
 
 impl ConsensusState {
